fix(billing): close plan-enforcement leaks across app mutations
- updateOrgPlan now triggers cascade downgrade (was bypass path)
- new orgs default to STARTER (was FREE, incoherent with fallback)
- extract enforceChannelDowngrade into lib/planEnforcement (shared)
- align channels.ts + ChannelsSettings filter to exclude banned too
- guard seedStandard and migration.migrateOrg with assertWithinLimit
